Added missing `.bbox()` calls to merge example
Hey there. I may be misunderstanding, so feel free to dismiss this, but the `merge` method example for `BBox` wasn't working. I looked at the source for a few minutes, and realized that the first two lines in the example were missing a call to `bbox`, since `merge` belongs to `BBox` instances, not `Rect` instances.

Thanks for making such a great lib. Really enjoying it so far :sparkles: