Sort out leftovers from the old fx module
This commit is mostly to clean up the implementation of the initAnimation
method of the Fx Module (issue #547). It also fixes these issues: #552, #582
and #584.

Here is a list of the changes and fixes that are made:
* Add a plot and array method to SVG.TextPath (issue #582)

* Make the method plot a getter when no parameter is passed for SVG.Polyline,
SVG.Polygon, SVG.Line, SVG.Path (this idea come from issue #547)

* Allow SVG.PointArray to be passed flat array

* Change the regexp SVG.PointArray use to parse string to allow more
flexibility in the way spaces and commas can be used to separate points
(something like this is now accepted: '1, 2, 3, 4')

* Allow plot to be called with 4 parameters when animating an SVG.Line

* Fix a bug in the plain morphing part of SVG.MorphObj that is in the FX module

* Relative value for SVG.Number are now calculated in its morph method (idea
from issue #547)

* Clean up the implementation of the initAnimation method of the FX module
(issues #547, #552, #584)