Allow 0 to be passed without unit
This commit finish fixing issue #552. It makes it unnecessary to specify a
unit with 0 when interacting with another SVG.Number that has a unit.

This was accomplished by modifying the operation methods of SVG.Number to
allow them to use the unit of the passed number when this number as no unit.