#805 Image memory leak fix
Removing image is not removing "load" and "error" events. (+2 events)
Removing image is not removing reference to image in DOM (+1 SVGImageElement detached dom)
This is fixed by passing this.node when registering "load" and "error" events instad of created img element.
Problem was that this.node could not be found in SVG.handlerMap in "off()" method.