Fixed an issue where a matrix string contained invalid floats
This commit fixes an issue which would cause errors in chrome when
a matrix like matrix(1, 0, 0, 1, 1e-90, 0) was output. This was
caused because all values in a matrix have to be valid float32's,
and therefore a 1e-90 could not be represented as a float32.

To fix this, we coax values smaller that 1e-27 to be represented
as zero. This avoids any potential issue due to this.